requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cssselect==1.3.0
cssutils==2.11.1
distro==1.9.0
et_xmlfile==2.0.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
griffe==1.7.3
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==6.0.0
mcp==1.10.1
more-itertools==10.7.0
numpy==2.3.1
oauth2client==4.1.3
oauthlib==3.3.1
openai==1.93.0
openai-agents==0.1.0
openpyxl==3.1.5
pandas==2.3.0
premailer==3.10.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyMuPDF==1.26.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.47.1
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
yagmail==0.15.293
